+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.5.1] - 2020-12-19
+
+### Fixed
+
+- Fixed terminal size detection on Windows https://github.com/willmcgugan/rich/issues/836
+- Fixed hex number highlighting
+
 ## [9.5.0] - 2020-12-18
 
 ### Changed